refactor: formatKRDate함수 utils 폴더로 이동
